Fix issue with extension not being found
Bun was failing to follow the redirects of your extension properly, so I added the slug infront of the id to skip the redirects. I also updated the display logic to filter out `null` items in the response so the rest of the items load even if one is null.