fix: Add another `defineItem` signature when `init` function is passed
https://github.com/wxt-dev/wxt/pull/1601 adjusted the typing of `defineItem` to
examine whether `defaultValue` or `fallback` were passed, and only then, remove
the `null` from the possible returned value. However, it did not accommodate
the case where an `init` function is passed to prevent a `null` return, which
was previously allowed.

Add a new type signature for this case, and add some tests for type
expectations to catch the issue.